refactor to isolate controller handlers from main app (to keep it clean and simple)
