+++ ws/data.js
@@ -1,0 +1,5 @@
+const clients = {};
+const sessions = {};
+const rooms = {};
+
+module.exports = { clients, sessions, rooms };
